update backward demodulation to reflect recent changes to Vampire
This commit removes the temporary hack, which was needed as Vampire's
backward demodulation rule inconsistenly saves its main premise as
second parent, in contrast to all other rules which save its main
premise as first parent.
A recent commit to Vampire fixes this inconsistency, so the hack is no
longer needed.