update literal flow reconstruction to support simultanous superposition
Recently, simultanous superposition was turned on as the default version
of superposition in Vampire. Until now, we did not support it, and
therefore raised an assertion (as simultanous superposition has the
same inference tag as standard superposition).

This commit adds support for simultanous superposition during
reconstruction of literal flows.

Reconstructing literal flows under the presence of both reorderings and
equality-rewritings is quite hard and probably expensive, so the
implementation uses educated guessing to reconstruct the literal flows
in the case of simultanous superposition. For the case of standard
superposition, where only one literal is rewritten, the reconstruction
is still guaranteed to be correct.