extend viz.js to use sufficient memory
This commit increases the memory available to graphviz inside
webassembly via viz.js from the default 16mb to 512mb.

It is motivated by the problem that 16mb is not enough to analyze bigger
proofs, which frequently causes crashes.

It seems to be possible to use even more than 512mb, but it is unclear
whether more than 1gb is possible and where the exact bound is and what
happens if one exceeds that bound. It therefore seems that 512mb is a
reasonable choice.